Merge "znmeb-is-collecting-commits" into "master"

+++ scripts/nginx.bash
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ -e /etc/nginx/sites-available/cropcompass ]
+then
+  echo `date` "nginx already configured - exiting"
+  exit
+fi
+
 echo `date` "Configuring nginx"
 sudo cp /vagrant/scripts/cropcompass.nginx /etc/nginx/sites-available/cropcompass
 sudo ln -s /etc/nginx/sites-available/cropcompass /etc/nginx/sites-enabled
