+++ scripts/uwsgi.bash
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ -e /etc/init/uwsgi.conf ]
+then
+  echo `date` "uWSGI already installed - exiting"
+  exit
+fi
+
 echo `date` "Installing uWSGI globally"
 sudo pip3 install uwsgi > ~/logs/uwsgi 2>&1
